refactor(engine): move local-write action specs into the shared engine package
src/mcp/local-write-tools.ts had zero root-specific dependencies (only
a generic JSON-value type) despite living in root src/ -- it builds
pure, deterministic LocalWriteActionSpecs (open_pr, file_issue,
apply_labels, post_eligibility_comment, create_branch, delete_branch,
generate_tests, follow_up_issue) with no DB/env/Cloudflare-binding
access, so it always belonged in the shared "brain" layer alongside
the rest of the portable engine, not root-only.

This lets packages/gittensory-miner's own real driving-loop entrypoint
(in progress) construct the exact same open_pr command gittensory's
MCP server would return via gittensory_open_pr, with zero network
round-trip and zero duplicated/drifting logic -- both consumers import
the same functions from one place.

src/mcp/local-write-tools.ts is now a thin re-export preserving every
existing import path (src/mcp/server.ts, src/review/fix-handoff-
render.ts, src/miner/soft-claim.ts) unchanged. Root gains
@jsonbored/gittensory-engine as an explicit dependency (the workspace
symlink already resolved it, but it was never declared).

test/unit/local-write-tools.test.ts adds the vi.mock source-redirect
this session's own established pattern uses whenever a root test
exercises logic that now lives behind the package-specifier re-export
-- without it, v8 coverage cannot attribute hits back to the engine's
.ts source (it resolves through dist/ instead), which would have
silently failed codecov/patch on the new engine file despite the
tests passing.
