+++ package.json
@@ -88,6 +88,7 @@
     "@asteasolutions/zod-to-openapi": "^8.5.0",
     "@cloudflare/puppeteer": "^1.1.0",
     "@hono/node-server": "^2.0.8",
+    "@jsonbored/gittensory-engine": "*",
     "@modelcontextprotocol/sdk": "1.29.0",
     "@octokit/core": "^7.0.6",
     "@opentelemetry/api": "^1.9.1",
